requirements.txt
pdfplumber
pandas
python-dateutil
regex
